tctl resource: `database object`
