+++ tool/tctl/common/collection.go
@@ -33,7 +33,6 @@ import (
 	"github.com/gravitational/teleport/api/defaults"
 	accessmonitoringrulesv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/accessmonitoringrules/v1"
 	crownjewelv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/crownjewel/v1"
-	dbobjectv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/dbobject/v1"
 	devicepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/devicetrust/v1"
 	healthcheckconfigv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/healthcheckconfig/v1"
 	loginrulepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/loginrule/v1"
@@ -53,7 +52,6 @@ import (
 	"github.com/gravitational/teleport/lib/sshutils"
 	"github.com/gravitational/teleport/lib/utils"
 	"github.com/gravitational/teleport/tool/common"
-	"github.com/gravitational/teleport/tool/tctl/common/databaseobject"
 	"github.com/gravitational/teleport/tool/tctl/common/loginrule"
 	"github.com/gravitational/teleport/tool/tctl/common/oktaassignment"
 	"github.com/gravitational/teleport/tool/tctl/common/resources"
@@ -923,32 +921,6 @@ func (c *samlIdPServiceProviderCollection) WriteText(w io.Writer, verbose bool) 
 	t := asciitable.MakeTable([]string{"Name"})
 	for _, serviceProvider := range c.serviceProviders {
 		t.AddRow([]string{serviceProvider.GetName()})
-	}
-	_, err := t.AsBuffer().WriteTo(w)
-	return trace.Wrap(err)
-}
-
-type databaseObjectCollection struct {
-	objects []*dbobjectv1.DatabaseObject
-}
-
-func (c *databaseObjectCollection) Resources() []types.Resource {
-	resources := make([]types.Resource, len(c.objects))
-	for i, b := range c.objects {
-		resources[i] = databaseobject.ProtoToResource(b)
-	}
-	return resources
-}
-
-func (c *databaseObjectCollection) WriteText(w io.Writer, verbose bool) error {
-	t := asciitable.MakeTable([]string{"Name", "Kind", "DB Service", "Protocol"})
-	for _, b := range c.objects {
-		t.AddRow([]string{
-			b.GetMetadata().GetName(),
-			fmt.Sprintf("%v", b.GetSpec().GetObjectKind()),
-			fmt.Sprintf("%v", b.GetSpec().GetDatabaseServiceName()),
-			fmt.Sprintf("%v", b.GetSpec().GetProtocol()),
-		})
 	}
 	_, err := t.AsBuffer().WriteTo(w)
 	return trace.Wrap(err)
